requirements.txt
# DURUM AI Agent - Python Dependencies

# Core
python-dotenv==1.0.0
requests==2.31.0

# APIs
pyairtable==2.3.3
facebook-business==19.0.0

# Slack
slack-sdk==3.26.2
slack-bolt==1.18.1

# Google Sheets
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# AI
anthropic==0.18.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Scheduling
APScheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
